msg: use uint32_t to represent nonce
so the type of nonce in messengers is consistent with
entity_addr_t::nonce. and it's more spatial efficient than uint64_t.

+++ src/msg/async/AsyncMessenger.cc
@@ -265,7 +265,7 @@ class C_handle_reap : public EventCallback {
  */
 
 AsyncMessenger::AsyncMessenger(CephContext *cct, entity_name_t name,
-                               string mname, uint64_t _nonce)
+                               string mname, uint32_t _nonce)
   : SimplePolicyMessenger(cct, name,mname, _nonce),
     dispatch_queue(cct, this, mname),
     lock("AsyncMessenger::lock"),
